move ReadsUtil to likelihood package
